# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Filament DB API",
    "description": "REST API for managing 3D printing filament profiles, nozzles, and printers. Supports filament inheritance (parent/variant), per-printer per-nozzle calibrations, INI profile export/import, and OpenPrintTag NFC binary generation.",
    "version": "1.8.4",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
    "/api/filaments": {
      "get": {
        "tags": ["Filaments"],
        "summary": "List filaments",
        "description": "Returns all non-deleted filaments, optionally filtered by type, vendor, or search term.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by filament type (e.g. PLA, PETG)"
          },
          {
            "name": "vendor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by vendor name"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Case-insensitive regex search on filament name"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of filament objects sorted by name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Filament" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Filaments"],
        "summary": "Create filament",
        "description": "Create a new filament profile. If parentId is provided, the filament becomes a color variant inheriting settings from its parent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FilamentInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created filament",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Filament" }
              }
            }
          },
          "400": {
            "description": "Invalid parent reference",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/filaments/import": {
      "post": {
        "tags": ["Filaments"],
        "summary": "Import filaments from INI",
        "description": "Imports filament profiles from an uploaded INI file. Creates new filaments or updates existing ones matched by name.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "INI file containing [filament:Name] sections"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "total": { "type": "integer" },
                    "created": { "type": "integer" },
                    "updated": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file provided or no profiles found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/filaments/import-atlas": {
      "post": {
        "tags": ["Filaments"],
        "summary": "Import from MongoDB Atlas",
        "description": "Connect to a remote MongoDB Atlas instance to list or import filaments. Send only `uri` to list available filaments; send `uri` + `filamentIds` to import selected ones.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["uri"],
                "properties": {
                  "uri": {
                    "type": "string",
                    "description": "MongoDB connection URI",
                    "example": "mongodb+srv://user:pass@cluster.mongodb.net/filament-db"
                  },
                  "filamentIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "IDs of filaments to import. Omit to list all available filaments."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of remote filaments (no filamentIds) or import result (with filamentIds)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "filaments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": { "type": "string" },
                              "name": { "type": "string" },
                              "vendor": { "type": "string" },
                              "type": { "type": "string" },
                              "color": { "type": "string" },
                              "parentId": { "type": "string", "nullable": true }
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "created": { "type": "integer" },
                        "updated": { "type": "integer" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing URI",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Connection failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/prusament/import": {
      "post": {
        "tags": ["Prusament"],
        "summary": "Import Prusament spool",
        "description": "Creates a filament from scraped Prusament data or adds a spool to an existing filament.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["spool"],
                "properties": {
                  "spool": { "$ref": "#/components/schemas/PrusamentSpool" },
                  "action": { "type": "string", "enum": ["create", "add-spool"], "description": "Create new filament or add spool to existing" },
                  "filamentId": { "type": "string", "nullable": true, "description": "Target filament ID (for add-spool action)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created filament", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Filament" } } } },
          "200": { "description": "Added spool to existing filament", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Filament" } } } },
          "400": { "description": "Missing spool data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/openprinttag/import": {
      "post": {
        "tags": ["OpenPrintTag"],
        "summary": "Import OpenPrintTag materials",
        "description": "Import selected OpenPrintTag materials into Filament DB. Materials are upserted by name (not name+vendor). If a filament with the same name exists under a different vendor, the import is skipped with an error.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["slugs"],
                "properties": { "slugs": { "type": "array", "items": { "type": "string" } } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResult" } } } },
          "400": { "description": "Missing slugs or database not loaded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
      "Temperatures": {
        "type": "object",
        "properties": {
          "nozzle": { "type": "number", "nullable": true, "description": "Nozzle temperature in °C" },
          "nozzleFirstLayer": { "type": "number", "nullable": true, "description": "First layer nozzle temperature" },
          "nozzleRangeMin": { "type": "number", "nullable": true, "description": "Minimum recommended nozzle temperature in °C" },
          "nozzleRangeMax": { "type": "number", "nullable": true, "description": "Maximum recommended nozzle temperature in °C" },
          "bed": { "type": "number", "nullable": true, "description": "Bed temperature in °C" },
          "bedFirstLayer": { "type": "number", "nullable": true, "description": "First layer bed temperature" },
          "standby": { "type": "number", "nullable": true, "description": "Standby/idle nozzle temperature in °C" }
        }
      },
      "Calibration": {
        "type": "object",
        "properties": {
          "printer": { "type": "string", "nullable": true, "description": "Printer ObjectId (null = default/any printer)" },
          "nozzle": { "type": "string", "description": "Nozzle ObjectId" },
          "extrusionMultiplier": { "type": "number", "nullable": true, "example": 0.95 },
          "maxVolumetricSpeed": { "type": "number", "nullable": true, "description": "mm³/s", "example": 15.0 },
          "pressureAdvance": { "type": "number", "nullable": true, "example": 0.053 },
          "retractLength": { "type": "number", "nullable": true, "description": "mm", "example": 0.8 },
          "retractSpeed": { "type": "number", "nullable": true, "description": "mm/s", "example": 35 },
          "retractLift": { "type": "number", "nullable": true, "description": "mm", "example": 0.2 }
        },
        "required": ["nozzle"]
      },
      "Preset": {
        "type": "object",
        "properties": {
          "label": { "type": "string", "description": "Preset name (e.g. Shore 85A)", "example": "Shore 85A" },
          "extrusionMultiplier": { "type": "number", "nullable": true },
          "temperatures": { "$ref": "#/components/schemas/Temperatures" }
        },
        "required": ["label"]
      },
      "Filament": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string", "example": "Prusament PLA Galaxy Black" },
          "vendor": { "type": "string", "example": "Prusament" },
          "type": { "type": "string", "example": "PLA" },
          "color": { "type": "string", "example": "#1a1a2e" },
          "cost": { "type": "number", "nullable": true, "description": "$/kg" },
          "density": { "type": "number", "nullable": true, "description": "g/cm³" },
          "diameter": { "type": "number", "default": 1.75, "description": "mm" },
          "temperatures": { "$ref": "#/components/schemas/Temperatures" },
          "maxVolumetricSpeed": { "type": "number", "nullable": true, "description": "mm³/s" },
          "compatibleNozzles": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of Nozzle ObjectIds"
          },
          "calibrations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Calibration" }
          },
          "presets": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Preset" }
          },
          "spoolWeight": { "type": "number", "nullable": true, "description": "Empty spool weight in grams" },
          "netFilamentWeight": { "type": "number", "nullable": true, "description": "Net filament weight in grams" },
          "totalWeight": { "type": "number", "nullable": true, "description": "Current total weight (spool + filament) in grams" },
          "instanceId": { "type": "string", "description": "Unique instance identifier (5-byte hex, e.g. 2acc21072a)", "example": "2acc21072a" },
          "dryingTemperature": { "type": "number", "nullable": true, "description": "Recommended drying temperature in °C" },
          "dryingTime": { "type": "number", "nullable": true, "description": "Recommended drying time in minutes" },
          "transmissionDistance": { "type": "number", "nullable": true, "description": "HueForge transmission distance" },
          "glassTempTransition": { "type": "number", "nullable": true, "description": "Glass transition temperature in °C (Tg)" },
          "heatDeflectionTemp": { "type": "number", "nullable": true, "description": "Heat deflection temperature in °C" },
          "shoreHardnessA": { "type": "number", "nullable": true, "description": "Shore A hardness (flexible/TPU filaments)" },
          "shoreHardnessD": { "type": "number", "nullable": true, "description": "Shore D hardness" },
          "shrinkageXY": { "type": "number", "nullable": true, "description": "XY-plane shrinkage compensation percentage" },
          "shrinkageZ": { "type": "number", "nullable": true, "description": "Z-axis shrinkage compensation percentage" },
          "minPrintSpeed": { "type": "number", "nullable": true, "description": "Minimum recommended print speed in mm/s" },
          "maxPrintSpeed": { "type": "number", "nullable": true, "description": "Maximum recommended print speed in mm/s" },
          "spoolType": { "type": "string", "nullable": true, "description": "Spool form factor (e.g. Standard, Refill, Cardboard)" },
          "colorName": { "type": "string", "nullable": true, "description": "Human-readable color name (e.g. Galaxy Black)" },
          "syncId": { "type": "string", "nullable": true, "description": "External sync identifier (e.g. PrusaSlicer preset name)" },
          "optTags": { "type": "array", "items": { "type": "integer" }, "description": "Optional tag flags (bit field values)" },
          "bedTypeTemps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "bedType": { "type": "string", "description": "Bed surface type (e.g. Textured PEI, Cool Plate, Engineering Plate)" },
                "temperature": { "type": "number", "nullable": true },
                "firstLayerTemperature": { "type": "number", "nullable": true }
              },
              "required": ["bedType"]
            },
            "description": "Per-bed-surface-type temperature overrides"
          },
          "abrasive": { "type": "boolean", "default": false, "description": "Whether the filament is abrasive" },
          "soluble": { "type": "boolean", "default": false, "description": "Whether the filament is soluble (support material)" },
          "spools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": { "type": "string" },
                "label": { "type": "string" },
                "totalWeight": { "type": "number", "nullable": true },
                "lotNumber": { "type": "string", "nullable": true, "description": "Manufacturer lot or batch number" },
                "purchaseDate": { "type": "string", "format": "date", "nullable": true, "description": "Date the spool was purchased" },
                "openedDate": { "type": "string", "format": "date", "nullable": true, "description": "Date the spool was opened" }
              }
            },
            "description": "Individual spool entries with weights"
          },
          "tdsUrl": { "type": "string", "nullable": true, "description": "Technical data sheet URL" },
          "inherits": { "type": "string", "nullable": true, "description": "Name of the slicer base profile this inherits from" },
          "parentId": { "type": "string", "nullable": true, "description": "Parent filament ID for color variants" },
          "settings": {
            "type": "object",
            "additionalProperties": { "type": "string", "nullable": true },
            "description": "Raw slicer key-value settings (e.g. extrusion_multiplier, start_filament_gcode)"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "vendor", "type"]
      },
      "FilamentInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "vendor": { "type": "string" },
          "type": { "type": "string" },
          "color": { "type": "string" },
          "cost": { "type": "number", "nullable": true },
          "density": { "type": "number", "nullable": true },
          "diameter": { "type": "number" },
          "temperatures": { "$ref": "#/components/schemas/Temperatures" },
          "maxVolumetricSpeed": { "type": "number", "nullable": true },
          "compatibleNozzles": {
            "type": "array",
            "items": { "type": "string" }
          },
          "calibrations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Calibration" }
          },
          "presets": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Preset" }
          },
          "spoolWeight": { "type": "number", "nullable": true },
          "netFilamentWeight": { "type": "number", "nullable": true },
          "totalWeight": { "type": "number", "nullable": true },
          "dryingTemperature": { "type": "number", "nullable": true },
          "dryingTime": { "type": "number", "nullable": true },
          "transmissionDistance": { "type": "number", "nullable": true },
          "glassTempTransition": { "type": "number", "nullable": true },
          "heatDeflectionTemp": { "type": "number", "nullable": true },
          "shoreHardnessA": { "type": "number", "nullable": true },
          "shoreHardnessD": { "type": "number", "nullable": true },
          "shrinkageXY": { "type": "number", "nullable": true },
          "shrinkageZ": { "type": "number", "nullable": true },
          "minPrintSpeed": { "type": "number", "nullable": true },
          "maxPrintSpeed": { "type": "number", "nullable": true },
          "spoolType": { "type": "string", "nullable": true },
          "colorName": { "type": "string", "nullable": true },
          "syncId": { "type": "string", "nullable": true },
          "optTags": { "type": "array", "items": { "type": "integer" } },
          "bedTypeTemps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "bedType": { "type": "string" },
                "temperature": { "type": "number", "nullable": true },
                "firstLayerTemperature": { "type": "number", "nullable": true }
              },
              "required": ["bedType"]
            }
          },
          "tdsUrl": { "type": "string", "nullable": true },
          "inherits": { "type": "string", "nullable": true },
          "parentId": { "type": "string", "nullable": true },
          "settings": {
            "type": "object",
            "additionalProperties": { "type": "string", "nullable": true }
          }
        },
        "required": ["name", "vendor", "type"]
      },
...
      "ImportResult": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "total": { "type": "integer" },
          "created": { "type": "integer" },
          "updated": { "type": "integer" },
          "skipped": { "type": "integer" }
        }
      },
      "PrusamentSpool": {
        "type": "object",
        "properties": {
          "spoolId": { "type": "string", "example": "c6974284da" },
          "productName": { "type": "string" },
          "material": { "type": "string" },
          "colorName": { "type": "string" },
          "colorHex": { "type": "string", "example": "#292929" },
          "diameter": { "type": "number", "example": 1.75 },
          "diameterAvg": { "type": "number" },
          "diameterStdDev": { "type": "number", "nullable": true },
          "ovality": { "type": "number" },
          "netWeight": { "type": "number", "description": "Net filament weight in grams" },
          "spoolWeight": { "type": "number", "description": "Empty spool weight in grams" },
          "totalWeight": { "type": "number" },
          "lengthMeters": { "type": "number" },
          "nozzleTempMin": { "type": "number" },
          "nozzleTempMax": { "type": "number" },
          "bedTempMin": { "type": "number" },
          "bedTempMax": { "type": "number" },
          "manufactureDate": { "type": "string" },
          "country": { "type": "string" },
          "goodsId": { "type": "integer" },
          "priceUsd": { "type": "number", "nullable": true },
          "priceEur": { "type": "number", "nullable": true },
          "photoUrl": { "type": "string" },
          "pageUrl": { "type": "string" }
        }
      }
    }
  }
}